synthetic_images/UNIT/requirements.txt
Pillow==5.4.1
numpy==1.16.2
tensorboardX==1.6
torch==1.0.1.post2
torchvision==0.2.2.post3
torchfile==0.1.0